+++ src/parser/enrichers/feat/_module.mjs
@@ -1,4 +1,4 @@
-export { default as AberrantDragonmark } from "./AbberantDragonmark.mjs";
+export { default as AbberantDragonmark } from "./AbberantDragonmark.mjs";
 export { default as AberrantFortitude } from "./AberrantFortitude.mjs";
 export { default as Alert } from "./Alert.mjs";
 export { default as BolsteringPerformance } from "./BolsteringPerformance.mjs";