+++ src/parser/enrichers/spell/AlustrielsMooncloak.mjs
@@ -1,0 +1,114 @@
+/* eslint-disable class-methods-use-this */
+import DDBEnricherData from "../data/DDBEnricherData.mjs";
+
+export default class AlustrielsMooncloak extends DDBEnricherData {
+
+  get type() {
+    return "utility";
+  }
+
+  get activity() {
+    return {
+      name: "Cast",
+    };
+  }
+
+  get addAutoAdditionalActivities() {
+    return false;
+  }
+
+  get additionalActivities() {
+    return [
+      {
+        constructor: {
+          name: "Liberation",
+          type: "utility",
+        },
+        build: {
+          generateConsumption: false,
+          noSpellslot: true,
+          generateAttack: false,
+          onsave: false,
+          noeffect: true,
+        },
+        overrides: {
+          activationType: "reaction",
+          overrideActivation: true,
+        },
+      },
+      {
+        constructor: {
+          name: "Respite",
+          type: "heal",
+        },
+        build: {
+          generateDamage: false,
+          generateHealing: true,
+          generateRange: true,
+          generateConsumption: true,
+          healingPart: DDBEnricherData.basicDamagePart({
+            number: 4,
+            denomination: 10,
+            bonus: "@mod",
+            type: "healing",
+          }),
+        },
+        overrides: {
+          activationType: "action",
+        },
+      },
+    ];
+  }
+
+  get effects() {
+    return [
+      {
+        activityMatch: "Cast",
+        name: "Within Moonlight",
+        options: {
+          durationSeconds: 60,
+          durationRounds: 10,
+        },
+        changes: [
+          DDBEnricherData.ChangeHelper.addChange("cold", 20, "system.traits.dr.value"),
+          DDBEnricherData.ChangeHelper.addChange("lightning", 20, "system.traits.dr.value"),
+          DDBEnricherData.ChangeHelper.addChange("radiant", 20, "system.traits.dr.value"),
+        ],
+        statuses: ["coverHalf"],
+        data: {
+          flags: {
+            dae: {
+              stackable: "noneNameOnly",
+              selfTarget: true,
+              selfTargetAlways: true,
+            },
+            ActiveAuras: {
+              aura: "Allies",
+              radius: "20",
+              isAura: true,
+              inactive: false,
+              hidden: false,
+              displayTemp: true,
+              ignoreSelf: false,
+              statuses: ["coverHalf"],
+            },
+          },
+        },
+        auraeffects: {
+          applyToSelf: true,
+          bestFormula: "",
+          canStack: false,
+          collisionTypes: ["move"],
+          combatOnly: false,
+          disableOnHidden: true,
+          distanceFormula: `20`,
+          disposition: 1,
+          evaluatePreApply: true,
+          overrideName: "",
+          script: "",
+        },
+      },
+    ];
+  }
+
+}